Cleanup: Removed unused variables and moved helper functions to utils.py
