+++ utils.py
@@ -1,3 +1,54 @@
+import json
+
+from bpy import context
+from pathlib import Path
+
+ADDON_FOLDER = Path(__file__).parent
+TALLY_FOLDER = ADDON_FOLDER / "tally_cache"
+
+# Create Folder for caching node tallies
+if not TALLY_FOLDER.exists():
+    TALLY_FOLDER.mkdir()
+
+
+def fetch_tally_path(tree_type):
+    return Path(TALLY_FOLDER, f'{tree_type}.json')    
+
+
+def fetch_user_prefs(attr_id=None):
+    prefs = context.preferences.addons[__package__].preferences
+    if attr_id is None:
+        return prefs
+    else:
+        return getattr(prefs, attr_id)
+
+
+def sort_enum_items(tree_type, items):
+    path = fetch_tally_path(tree_type)
+    if path.exists():
+        with open(path, "r") as f:
+            tally_dict = json.load(f)
+
+        items.sort(key=lambda x : tally_dict.get(x[0], 0), reverse=True)
+
+
+def update_tally(context, entry):
+    prefs = fetch_user_prefs() 
+    tree_type = context.space_data.tree_type
+
+    path = fetch_tally_path(tree_type)
+    if path.exists():
+        with open(path, "r") as f:
+            tally_dict = json.load(f)
+    else:
+        tally_dict = {}
+
+    tally_dict[entry] = min(tally_dict.get(entry, 0) + 1, prefs.tally_weight)
+
+    with open(path, "w") as f:
+        json.dump(tally_dict, f, indent=4)
+
+
 def in_nodegroup(context):
     current_tree = context.space_data.edit_tree
     node_groups = context.blend_data.node_groups
